+++ packages/terra-form-checkbox/src/index.js
@@ -1,0 +1,5 @@
+import Checkbox from './Checkbox';
+import CheckboxField from './CheckboxField';
+
+export default Checkbox;
+export { CheckboxField };