+++ packages/terra-form-textarea/src/index.js
@@ -1,0 +1,5 @@
+import Textarea from './Textarea';
+import TextareaField from './TextareaField';
+
+export default Textarea;
+export { TextareaField };