+++ packages/terra-core-docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Changed
   * Changed `terra responsive element` docs to include reflow details.
   * Updated `terra-alert` example to use `titleID` prop.
+  * Updated usage documentation for terra-form-checkbox, terra-form-input, terra-form-radio & terra-form-textarea.
 
 ## 1.66.0 - (March 12, 2024)